requirements.txt
appdirs==1.4.3
docutils==0.13.1
lockfile==0.12.2
packaging==16.8
pyparsing==2.2.0
python-daemon==2.1.2
requests==2.14.2
six==1.10.0
sseclient==0.0.18
